Loggers defer configuration until used
Previously New returned a logger that used the default options. To configure
a logger with a different output format or destination it was necessary to
use NewWithOptions. However it is awkward to call NewWithOptions in every
package or type that uses logging and busywork to pass a hierarchy
of child loggers down to each package or type before using it.

This change makes loggers defer applying their configuration until they
are used. The logger is instantiated when either Info, Error or Enabled is
called. At that point the logger will read and use any options set from a
prior call to UseOptions. Child loggers derived from other loggers will base
their configuration on their parent's which may trigger the parent to
instantiate itself.